+++ AmuEngine/AmuEngine/Sources/Map/Grid.h
@@ -9,12 +9,12 @@ class Grid
 public:
     std::vector<float> vertices;
     Grid() : gridVAO(0), gridVBO(0), vertices({}) {};
-    Grid(float gridWidth_, float gridHeight_, float gridSpacing_) : gridVAO(0), gridVBO(0), vertices({}) {
+    Grid(int gridWidth_, int gridHeight_, int gridSpacing_) : gridVAO(0), gridVBO(0), vertices({}) {
         SetupGrid(gridWidth_, gridHeight_, gridSpacing_);
     };
-    void SetupGrid(float gridWidth, float gridHeight, float gridSpacing);
+    void SetupGrid(int gridWidth, int gridHeight, int gridSpacing);
 
-    std::vector<float> GenerateGridVertices(float gridWidth, float gridHeight, float gridSpacing);
+    std::vector<float> GenerateGridVertices(int gridWidth, int gridHeight, int gridSpacing);
 
     void RenderGrid();
 };
